requirements.txt
# Core
python-dotenv>=1.0.0
numpy>=1.24.0

# Data processing
tqdm>=4.66.0

# Embeddings (local, no API)
sentence-transformers>=2.2.2

# Vector database
faiss-cpu>=1.7.4

# Backend API
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Text-to-Speech (simple & stable)
gTTS>=2.5.1